+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,17 @@
 # examples
 
+## 0.3.0
+
+### Minor Changes
+
+- b86b252: add getProperty method on the device to query for a specific property value
+
+### Patch Changes
+
+- Updated dependencies [b86b252]
+  - @ecoflow-api/rest-client@0.3.0
+  - @ecoflow-api/schemas@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes